requirements.txt
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
click==8.1.3
fastapi==0.75.2
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
Mako==1.2.0
MarkupSafe==2.1.1
pydantic==1.9.0
python-dotenv==0.20.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.17.1
typing-extensions==4.2.0
uvicorn==0.17.6
zipp==3.8.0
psycopg2==2.9.1
